configure.ac changes to support RTEMS 6
- Disabling GeSys and rtems-gdb-stub

- Added detection for BSD vs. LEGACY network stacks

GeSys is going to be replaced by a new init system, since it's just
easier to work from scratch, especially since we're going to try to use
the RTEMS RTL (runtime linker) instead of Cexp w/pmelf and such.

rtems-gdb-stub is replaced by libdebugger in RTEMS 6 (and 5??), but it
does not support m68k. At least initially, we're going to be missing
m68k debugging support. Eventually I'd like to add it to libdebugger.